# Optimize NCHW2NHWC on GPU