--- a/caffe2/operators/order_switch_ops.cc
+++ b/caffe2/operators/order_switch_ops.cc
@@ -1,81 +1,8 @@
 #include "caffe2/operators/order_switch_ops.h"
 
+#include <string>
+
 namespace caffe2 {
-
-template <>
-bool NHWC2NCHWOp<float, CPUContext>::RunOnDevice() {
-  auto& X = Input(0);
-  auto* Y = Output(0);
-
-  auto ndim = X.ndim();
-  CAFFE_ENFORCE_GE(ndim, 3);
-  const int N = X.dim32(0), C = X.dim32(ndim - 1);
-  vector<int64_t> Y_dims(ndim);
-  Y_dims[0] = N;
-  Y_dims[1] = C;
-  int image_size = 1;
-  for (auto i = 2; i < ndim; ++i) {
-    Y_dims[i] = X.dim32(i - 1);
-    image_size *= Y_dims[i];
-  }
-  Y->Resize(Y_dims);
-  if (X.size() <= 0) {
-    return true;
-  }
-
-  const float* Xdata = X.data<float>();
-  float* Ydata = Y->mutable_data<float>();
-  std::array<int, 2> dims = {image_size, C};
-  std::array<int, 2> axes = {1, 0};
-  for (int n = 0; n < N; ++n) {
-    math::Transpose(
-        2,
-        dims.data(),
-        axes.data(),
-        Xdata + n * image_size * C,
-        Ydata + n * image_size * C,
-        &context_);
-  }
-  return true;
-}
-
-template <>
-bool NCHW2NHWCOp<float, CPUContext>::RunOnDevice() {
-  auto& X = Input(0);
-  auto* Y = Output(0);
-
-  auto ndim = X.ndim();
-  CAFFE_ENFORCE_GE(X.ndim(), 3);
-  const int N = X.dim32(0), C = X.dim32(1);
-  vector<int64_t> Y_dims(ndim);
-  Y_dims[0] = N;
-  int image_size = 1;
-  for (auto i = 1; i < ndim - 1; ++i) {
-    Y_dims[i] = X.dim32(i + 1);
-    image_size *= Y_dims[i];
-  }
-  Y_dims[ndim - 1] = C;
-  Y->Resize(Y_dims);
-  if (X.size() <= 0) {
-    return true;
-  }
-
-  const float* Xdata = X.data<float>();
-  float* Ydata = Y->mutable_data<float>();
-  std::array<int, 2> dims = {C, image_size};
-  std::array<int, 2> axes = {1, 0};
-  for (int n = 0; n < N; ++n) {
-    math::Transpose(
-        2,
-        dims.data(),
-        axes.data(),
-        Xdata + n * image_size * C,
-        Ydata + n * image_size * C,
-        &context_);
-  }
-  return true;
-}
-
 
 REGISTER_CPU_OPERATOR(NHWC2NCHW, NHWC2NCHWOp<float, CPUContext>);
 REGISTER_CPU_OPERATOR(NCHW2NHWC, NCHW2NHWCOp<float, CPUContext>);
@@ -84,10 +11,10 @@ OPERATOR_SCHEMA(NHWC2NCHW)
     .NumInputs(1)
     .NumOutputs(1)
     .TensorInferenceFunction([](const OperatorDef& /*unused*/ /*def*/,
-                                const vector<TensorShape>& in) {
+                                const std::vector<TensorShape>& in) {
       CAFFE_ENFORCE_GE(
           in[0].dims_size(), 3, "Input for NHWC2NCHW must be >= 3 dimensional");
-      vector<TensorShape> out(1);
+      std::vector<TensorShape> out(1);
       out[0].add_dims(in[0].dims(0));
       out[0].add_dims(in[0].dims(in[0].dims_size() - 1));
       for (auto i = 1; i < in[0].dims_size() - 1; ++i) {
@@ -109,10 +36,10 @@ OPERATOR_SCHEMA(NCHW2NHWC)
     .NumInputs(1)
     .NumOutputs(1)
     .TensorInferenceFunction([](const OperatorDef& /*unused*/ /*def*/,
-                                const vector<TensorShape>& in) {
+                                const std::vector<TensorShape>& in) {
       CAFFE_ENFORCE_GE(
           in[0].dims_size(), 3, "Input for NCHW2NHWC must be >= 3 dimensional");
-      vector<TensorShape> out(1);
+      std::vector<TensorShape> out(1);
       out[0].add_dims(in[0].dims(0));
       for (auto i = 2; i < in[0].dims_size(); ++i) {
         out[0].add_dims(in[0].dims(i));
@@ -130,25 +57,33 @@ k=1, 2, and 3.
     .Input(0, "data", "The input data (Tensor) in the NCHW order.")
     .Output(0, "output", "The output tensor (Tensor) in the NHWC order.");
 
+namespace {
+
 class GetNHWC2NCHWGradient : public GradientMakerBase {
   using GradientMakerBase::GradientMakerBase;
-  vector<OperatorDef> GetGradientDefs() override {
+  std::vector<OperatorDef> GetGradientDefs() override {
     return SingleGradientDef(
-        "NCHW2NHWC", "",
-        vector<string>{GO(0)},
-        vector<string>{GI(0)});
+        "NCHW2NHWC",
+        "",
+        std::vector<std::string>{GO(0)},
+        std::vector<std::string>{GI(0)});
   }
 };
-REGISTER_GRADIENT(NHWC2NCHW, GetNHWC2NCHWGradient);
 
 class GetNCHW2NHWCGradient : public GradientMakerBase {
   using GradientMakerBase::GradientMakerBase;
-  vector<OperatorDef> GetGradientDefs() override {
+  std::vector<OperatorDef> GetGradientDefs() override {
     return SingleGradientDef(
-        "NHWC2NCHW", "",
-        vector<string>{GO(0)},
-        vector<string>{GI(0)});
+        "NHWC2NCHW",
+        "",
+        std::vector<std::string>{GO(0)},
+        std::vector<std::string>{GI(0)});
   }
 };
+
+} // namespace
+
+REGISTER_GRADIENT(NHWC2NCHW, GetNHWC2NCHWGradient);
 REGISTER_GRADIENT(NCHW2NHWC, GetNCHW2NHWCGradient);
-}  // namespace caffe2
+
+} // namespace caffe2