--- a/caffe2/utils/math_utils.cc
+++ b/caffe2/utils/math_utils.cc
@@ -270,6 +270,18 @@ bool IsBothEndsBroadcastBinaryOp(
   return true;
 }
 
+bool IsBatchTranspose2D(const int ndim, const int* axes) {
+  if (ndim < 2) {
+    return false;
+  }
+  for (int i = 0; i < ndim - 2; ++i) {
+    if (axes[i] != i) {
+      return false;
+    }
+  }
+  return axes[ndim - 2] == ndim - 1 && axes[ndim - 1] == ndim - 2;
+}
+
 void ComputeTransposeAxesForReduceOp(
     const int num_dims,
     const int num_reduce_axes,