Summary: Optimize NCHW2NHWC on GPU

Differential Revision: D10481163