@BIT-silence 2018-10-21
benchmark:
N = 128, C = 256, H = 152, W = 152
NCNW2NHWC: 401.25ms -> 35.58ms
NHWC2NCHW: 339.50ms -> 37.44ms