@BIT-silence 2018-10-21
2D Transpose,
R = 10000, C = 10000,
58ms -> 5.7ms
CuDNN: 20.78ms